dom/presentation/tests/mochitest/test_presentation_sender_on_terminate_request.html

<!DOCTYPE HTML>
<html>
<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<head>
  <meta charset="utf-8">
  <title>Test onTerminateRequest at sender side</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1276378">Test onTerminateRequest at sender side</a>
<script type="application/javascript">

"use strict";

var gScript = SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL("PresentationSessionChromeScript.js"));
var request;
var connection;

function testSetup() {
  return new Promise(function(aResolve, aReject) {
    request = new PresentationRequest("http://example.com");

    request.getAvailability().then(
      function(aAvailability) {
        is(aAvailability.value, false, "Sender: should have no available device after setup");
        aAvailability.onchange = function() {
          aAvailability.onchange = null;
          ok(aAvailability.value, "Device should be available.");
          aResolve();
        };

        gScript.sendAsyncMessage("trigger-device-add");
      },
      function(aError) {
        ok(false, "Error occurred when getting availability: " + aError);
        teardown();
        aReject();
      }
    );
  });
}

function testStartConnection() {
  return new Promise(function(aResolve, aReject) {
    gScript.addMessageListener("device-prompt", function devicePromptHandler() {
      gScript.removeMessageListener("device-prompt", devicePromptHandler);
      info("Device prompt is triggered.");
      gScript.sendAsyncMessage("trigger-device-prompt-select");
    });

    gScript.addMessageListener("control-channel-established", function controlChannelEstablishedHandler() {
      gScript.removeMessageListener("control-channel-established", controlChannelEstablishedHandler);
      info("A control channel is established.");
      gScript.sendAsyncMessage("trigger-control-channel-open");
    });

    gScript.addMessageListener("control-channel-opened", function controlChannelOpenedHandler(aReason) {
      gScript.removeMessageListener("control-channel-opened", controlChannelOpenedHandler);
      info("The control channel is opened.");
    });

    gScript.addMessageListener("control-channel-closed", function controlChannelClosedHandler(aReason) {
      gScript.removeMessageListener("control-channel-closed", controlChannelClosedHandler);
      info("The control channel is closed. " + aReason);
    });

    gScript.addMessageListener("offer-sent", function offerSentHandler(aIsValid) {
      gScript.removeMessageListener("offer-sent", offerSentHandler);
      ok(aIsValid, "A valid offer is sent out.");
      gScript.sendAsyncMessage("trigger-incoming-transport");
    });

    gScript.addMessageListener("answer-received", function answerReceivedHandler() {
      gScript.removeMessageListener("answer-received", answerReceivedHandler);
      info("An answer is received.");
    });

    gScript.addMessageListener("data-transport-initialized", function dataTransportInitializedHandler() {
      gScript.removeMessageListener("data-transport-initialized", dataTransportInitializedHandler);
      info("Data transport channel is initialized.");
      gScript.sendAsyncMessage("trigger-incoming-answer");
    });

    gScript.addMessageListener("data-transport-notification-enabled", function dataTransportNotificationEnabledHandler() {
      gScript.removeMessageListener("data-transport-notification-enabled", dataTransportNotificationEnabledHandler);
      info("Data notification is enabled for data transport channel.");
    });

    var connectionFromEvent;
    request.onconnectionavailable = function(aEvent) {
      request.onconnectionavailable = null;
      connectionFromEvent = aEvent.connection;
      ok(connectionFromEvent, "|connectionavailable| event is fired with a connection.");

      if (connection) {
        is(connection, connectionFromEvent, "The connection from promise and the one from |connectionavailable| event should be the same.");
      }
    };

    request.start().then(
      function(aConnection) {
        connection = aConnection;
        ok(connection, "Connection should be available.");
        ok(connection.id, "Connection ID should be set.");
        is(connection.state, "connecting", "The initial state should be connecting.");

        if (connectionFromEvent) {
          is(connection, connectionFromEvent, "The connection from promise and the one from |connectionavailable| event should be the same.");
        }
        connection.onconnect = function() {
          connection.onconnect = null;
          is(connection.state, "connected", "Connection should be connected.");
          aResolve();
        };
      },
      function(aError) {
        ok(false, "Error occurred when establishing a connection: " + aError);
        teardown();
        aReject();
      }
    );
  });
}

function testOnTerminateRequest() {
  return new Promise(function(aResolve, aReject) {
    gScript.addMessageListener("control-channel-opened", function controlChannelOpenedHandler(aReason) {
      gScript.removeMessageListener("control-channel-opened", controlChannelOpenedHandler);
      info("The control channel is opened.");
    });

    gScript.addMessageListener("control-channel-closed", function controlChannelClosedHandler(aReason) {
      gScript.removeMessageListener("control-channel-closed", controlChannelClosedHandler);
      info("The control channel is closed. " + aReason);
    });

    gScript.addMessageListener("data-transport-closed", function dataTransportClosedHandler(aReason) {
      gScript.removeMessageListener("data-transport-closed", dataTransportClosedHandler);
      info("The data transport is closed. " + aReason);
    });

    connection.onterminate = function() {
      connection.onterminate = null;
      is(connection.state, "terminated", "Connection should be closed.");
      aResolve();
    };

    gScript.sendAsyncMessage("trigger-incoming-terminate-request");
    gScript.sendAsyncMessage("trigger-control-channel-open");
  });
}

function teardown() {
  gScript.addMessageListener("teardown-complete", function teardownCompleteHandler() {
    gScript.removeMessageListener("teardown-complete", teardownCompleteHandler);
    gScript.destroy();
    SimpleTest.finish();
  });

  gScript.sendAsyncMessage("teardown");
}

function runTests() {
  ok(window.PresentationRequest, "PresentationRequest should be available.");

  testSetup().
  then(testStartConnection).
  then(testOnTerminateRequest).
  then(teardown);
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPermissions([
  {type: "presentation-device-manage", allow: false, context: document},
], function() {
  SpecialPowers.pushPrefEnv({ "set": [["dom.presentation.enabled", true],
                                      ["dom.presentation.controller.enabled", true],
                                      ["dom.presentation.receiver.enabled", false],
                                      ["dom.presentation.session_transport.data_channel.enable", false]]},
                            runTests);
});

</script>
</body>
</html>
